Split up IO module (#203)
* User-friendly error when setting manually ds.sections

Solves #200

* User-friendly error when setting manually ds.sections

Solves #200

* Revert accidental double implementation of sections.setter

* Update documentation of datastore class

* Removed numbagg from dependencies

llvmlite is a pain to install with pip. Alternatively, it could be installed with conda

* Explicitly pass sections to more functions

* Format

* Moving away from calibrate_double_ended() inplace

self.update(out) is still in place at the end

* Refactored calibrate_double_ended covariance calculation

* Parallel testing

* refactored variance functions

* Include pyupgrade flag of ruff

* Refactored single ended routine p_val to params and moved averaging routines

* Update .gitignore

* Refactored calibration routines

In preparation of the accessor

* Please ruff

* Refactored Datastore class

* Updated sections example notebook

* Updated example notebooks

* Add api to docs

* Unfortenate formatting of notebooks

* Split and move IO routines.

* Filter out some common warnings.

---------
