+++ src/dtscalibration/__init__.py
@@ -5,12 +5,12 @@ from dtscalibration.datastore_utils import get_netcdf_encoding
 from dtscalibration.datastore_utils import merge_double_ended
 from dtscalibration.datastore_utils import shift_double_ended
 from dtscalibration.datastore_utils import suggest_cable_shift_double_ended
-from dtscalibration.io import open_datastore
-from dtscalibration.io import open_mf_datastore
-from dtscalibration.io import read_apsensing_files
-from dtscalibration.io import read_sensornet_files
-from dtscalibration.io import read_sensortran_files
-from dtscalibration.io import read_silixa_files
+from dtscalibration.io.apsensing import read_apsensing_files
+from dtscalibration.io.datastore import open_datastore
+from dtscalibration.io.datastore import open_mf_datastore
+from dtscalibration.io.sensornet import read_sensornet_files
+from dtscalibration.io.sensortran import read_sensortran_files
+from dtscalibration.io.silixa import read_silixa_files
 from dtscalibration.plot import plot_accuracy
 from dtscalibration.plot import plot_location_residuals_double_ended
 from dtscalibration.plot import plot_residuals_reference_sections
