+++ tests/test_dtscalibration.py
@@ -1,4 +1,5 @@
 import os
+import warnings
 
 import numpy as np
 import pytest
@@ -40,7 +41,11 @@ else:
 def assert_almost_equal_verbose(actual, desired, verbose=False, **kwargs):
     """Print the actual precision decimals"""
     err = np.abs(actual - desired).max()
-    dec = -np.ceil(np.log10(err))
+
+    with warnings.catch_warnings():
+        # Supress zero division warnings
+        warnings.filterwarnings("ignore", message="divide by zero encountered in log10")
+        dec = -np.ceil(np.log10(err))
 
     if not (np.isfinite(dec)):
         dec = 18.0
